requirements.txt
annotated-types==0.7.0
appwrite==18.0.0
Authlib==1.7.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
gunicorn==25.3.0
idna==3.12
itsdangerous==2.2.0
Jinja2==3.1.6
joserfc==1.6.4
MarkupSafe==3.0.3
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
redis==4.5.4
